refactor of model_generator
