--- nn_meter/builder/model_generator/generator/__init__.py
@@ -1,3 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT license.
-from generator import Generator
