--- nn_meter/builder/model_generator/generator/networks/ops/__init__.py
@@ -1,2 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT license.
