+++ nn_meter/builder/model_generator/ruletest_model.py
@@ -3,6 +3,7 @@
 import tensorflow as tf
 
 from . import operators
+from .utils import get_op_output_shape, get_op_is_two_inputs
 from nn_meter.builder.utils import get_inputs_by_shapes
 
 
@@ -41,7 +42,10 @@ class TwoOpModel(tf.keras.Model):
 
 
 def get_operator_by_name(name, input_shape, config = None):
-    return getattr(operators, name)(input_shape, config)
+    operator = getattr(operators, name)(input_shape, config)
+    output_shape = get_op_output_shape(name, input_shape)
+    op_is_two_inputs = get_op_is_two_inputs(name)
+    return operator, output_shape, op_is_two_inputs
 
 
 def get_ruletest_model(op1, op2, input_shape, config):
